🎉 Initialize project
This project is based on another project + a fork of the project.
- https://github.com/corbosman/node-red-contrib-timerswitch
- fork: https://github.com/fbuys/node-red-contrib-timerswitch

I decided to create new project because my intention is to refactor it
to suite my needs. It would also not be accurate to call it a fork,
because I don't intend to merge into or stay aligned with the upstream
branch of the fork.